backend/requirements.txt
Django>=4.2.7,<4.3
django-ninja>=1.0.1,<1.1
psycopg[binary]>=3.1.8,<3.2
wait-for-it>=2.2.2,<2.3
pyjwt[crypto]>=2.8.0,<2.9
python-decouple>=3.7,<3.8
